# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the API for the new version of the XCCup.net.\n",
    "version": "1.0.0",
    "title": "XCCup",
    "termsOfService": "https://www.xccup.net",
    "contact": {
      "email": "admin@xccup.net"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost",
  "basePath": "/",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/flight": {
      "get": {
        "tags": ["flight"],
        "summary": "Get all flights of current season",
        "description": "Returns all flights of the current season",
        "operationId": "getFlightsOfCurrentSeason",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BasicFlight"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["flight"],
        "summary": "Add a new flight to the database",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Flight object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Flight"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": ["flight"],
        "summary": "Update an existing flight",
        "operationId": "updateFlight",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Flight object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Flight"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Flight not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/flight/{flightId}": {
      "get": {
        "tags": ["flight"],
        "summary": "Find flight by ID",
        "description": "Returns a single flight",
        "operationId": "getFlightById",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "description": "ID of flight to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Flight"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Flight not found"
          }
        }
      },
      "delete": {
        "tags": ["flight"],
        "summary": "Deletes a flight",
        "operationId": "deleteFlight",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "description": "Flight id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Flight not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BasicFlight": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "pilot": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "photoUrl": {
              "type": "string"
            }
          }
        },
        "clubName": {
          "type": "string"
        },
        "teamName": {
          "type": "string"
        },
        "aircraft": {
          "$ref": "#/definitions/Aircraft"
        },
        "result" : {
          "type" : "object",
          "properties" : {
            "status" : {
              "type" : "string",
              "enum" : [ "Nicht in Wertung", "In Wertung", "Flugbuch", "In Bearbeitung" ]
            },
            "points" : {
              "type" : "number"
            },
            "distance" : {
              "type" : "number"
            },
            "type" : {
              "type" : "string"
            },
            "turnpoints" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        },
        "submitted": {
          "type": "string",
          "format": "date-time"
        },
        "igcUrl": {
          "type": "string"
        }
      }
    },
    "Flight": {
      "type": "object",
      "properties": {
        "basic": {
          "$ref": "#/definitions/BasicFlight"
        },
        "date": { "type": "string", "format": "date-time" },
        "pilot": {
          "type": "object",
          "properties": {
            "userId": "UUID",
            "name": "string",
            "club": {
              "type": "object",
              "items": {
                "name": "string",
                "id": "UUID"
              }
            },
            "team": {
              "type": "object",
              "items": {
                "name": "string",
                "id": "UUID"
              }
            }
          }
        },
        "details": {
          "type": "object",
          "items": {
            "distance": "integer",
            "taskType": "string",
            "points": "integer",
            "takeoff": {
              "type": "object",
              "items": {
                "name": "string",
                "id": "UUID"
              }
            },
            "landing": {
              "type": "object",
              "items": {
                "name": "string",
                "id": "UUID"
              }
            },
            "takeoffTime": { "type": "string", "format": "date-time" },
            "landingTime": { "type": "string", "format": "date-time" },
            "glider": "string",
            "rankingClass": "string",
            "flightDuration": "string",
            "rankingState": "string",
            "maxAlt": "string",
            "minAlt": "string",
            "maxClimbRate": "string",
            "maxSinkRate": "string",
            "averageSpeed": "string",
            "taskSpeed": "string",
            "uploadDate": { "type": "string", "format": "date-time" },
            "igcUrl": "url",
            "flightReport": "string"
          }
        },
        "images": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "url": {
                "type": "url"
              },
              "description": {
                "type": "string"
              },
              "id": {
                "type": "UUID"
              }
            }
          }
        },
        "airbuddies": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "flightId": {
                "type": "integer"
              },
              "userId": {
                "type": "UUID"
              },
              "name": {
                "type": "string"
              },
              "sharedTime": {
                "type": "integer"
              }
            }
          }
        },
        "comments": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "userId": {
                "type": "UUID"
              },
              "text": "string",
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "userAvatar": "url"
            }
          }
        },
        "fixes": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "latitute": {
                "type": "string"
              },
              "longitute": {
                "type": "string"
              },
              "elevation": {
                "type": "integer"
              },
              "gpsAltitude": {
                "type": "integer"
              },
              "baroAltitude": {
                "type": "integer"
              }
            }
          }
        }
      }
    },
...
    "Aircraft": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string"
        },
        "model": {
          "type": "string"
        },
        "rankingClass": {
          "type": "string"
        }
      }
    },
...
  }
}